// apps/api/src/setting/setting.constant.ts
export enum CountrySortProperty {
  Id = 'id',
  Name = 'name',
  Code = 'code',
  DialCode = 'dial_code',
  CreatedAt = 'created_at',
  UpdatedAt = 'updated_at',
  DeletedAt = 'deleted_at',
}

export enum SurfboardSortProperty {
  Id = 'id',
  Name = 'name',
  CreatedAt = 'created_at',
  UpdatedAt = 'updated_at',
  DeletedAt = 'deleted_at',
}